+++ packages/vuepress-plugin-export-pdf-v2/src/commands/index.ts
@@ -1,1 +1,2 @@
 export * from "../index";
+export * from "./press-export-pdf";
