+++ packages/vuepress-plugin-export-pdf/src/commands/index.ts
@@ -1,1 +1,2 @@
 export * from "../index";
+export * from "./press-export-pdf";
