--- packages/vuepress-plugin-export-pdf/src/configs/index.ts
@@ -1,5 +1,0 @@
-export * from "./resolveUserConfigConventionalPath";
-export * from "./resolveUserConfigPath";
-export * from "./defineUserConfig";
-export * from "./userConfigTypes";
-export * from "./loadModule";
