--- packages/vuepress-plugin-export-pdf/src/dev-server/index.ts
@@ -1,2 +1,0 @@
-export * from "./vuePressServer";
-export * from "./serverApp";
