+++ packages/vuepress-plugin-export-pdf/src/index.ts
@@ -1,6 +1,4 @@
-export * from "./runner";
-export * from "./utils";
-export * from "./configs";
-export * from "./systemInfo";
-export * from "./dev-server";
+export * from "./types";
+export * from "./serverApp";
 export * from "./extendCli";
+export * from "./defineUserConfig";
