--- packages/vuepress-plugin-export-pdf/src/utils/index.ts
@@ -1,7 +1,0 @@
-export * from "./timeTransformer";
-export * from "./wrapCommand";
-export * from "./generatePdf";
-export * from "./filterRoute";
-export * from "./checkEnv";
-export * from "./mergePDF";
-export * from "./progress";
