--- packages/vuepress-plugin-export-pdf/src/utils/progress.ts
@@ -1,7 +1,0 @@
-import cliProgress from "cli-progress";
-
-export const singleProgressBar = (format = "Progress |" + "{bar}" + "| {percentage} % || {value} / {total}", preset = cliProgress.Presets.shades_classic) => {
-	return new cliProgress.Bar({
-		format,
-	}, preset);
-};
